services/microservicio-ia/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.7
google-generativeai==0.3.1
structlog==23.2.0
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3
